+++ CHANGES.rst
@@ -43,6 +43,7 @@ in progress
 - Refactoring: Add ``eskema.util`` folder instead of single ``util.py``
 - UX: Add ``eskema info`` subcommand
 - UX: Improve CLI help
+- Format: Add support for reading compressed files in Gzip format
 
 
 2023-01-xx 0.0.0